생성 : fund신청, 취소, 사용자조회, 관리자조회

+++ Back/neighfund/src/main/java/org/project/neighfund/application/fund/dto/FundResponseDto.java
@@ -35,7 +35,7 @@ public class FundResponseDto {
 
     private Integer progressRate;  //달성률
     private Integer targetAmount;  //목표금액
-    private Integer currentAmount;  //현재 모인 금액
+    private Long currentAmount;  //현재 모인 금액
     private Integer currentParticipants = 0;  //현재참여자
     private LocalDateTime deadline;  //마감일
     private String hashTags;
